New feature: 3266524: Return terms from non stored fields New feature: 3266520: Cache management in Web UI

+++ src/java/com/jaeksoft/searchlib/crawler/web/screenshot/ScreenshotManager.java
@@ -192,7 +192,9 @@ public class ScreenshotManager implements PropertyItemListener {
 		for (String u : urlList) {
 			try {
 				URL url = new URL(u);
-				getPngFile(url).delete();
+				File f = getPngFile(url);
+				if (f != null)
+					getPngFile(url).delete();
 			} catch (MalformedURLException e) {
 				Logging.logger.warn(e);
 			}
